browserify - move core/createPackageDataStream -> browserify
