+++ packages/browserify/src/index.js
@@ -4,9 +4,10 @@ const clone = require('clone')
 const mkdirp = require('mkdirp')
 const through = require('through2').obj
 const mergeDeep = require('merge-deep')
-const { generatePrelude, createPackageDataStream } = require('lavamoat-core')
+const { generatePrelude } = require('lavamoat-core')
 const jsonStringify = require('json-stable-stringify')
 const { createModuleInspectorSpy } = require('./createModuleInspectorSpy.js')
+const { createPackageDataStream } = require('./createPackageDataStream.js')
 const createCustomPack = require('./createCustomPack')
 const { createSesWorkaroundsTransform } = require('./sesTransforms')
