+++ packages/core/package.json
@@ -15,8 +15,7 @@
     "json-stable-stringify": "^1.0.1",
     "lavamoat-tofu": "^5.0.0",
     "resolve": "^1.15.1",
-    "ses": "^0.7.7",
-    "through2": "^3.0.1"
+    "ses": "^0.7.7"
   },
   "devDependencies": {
     "merge-deep": "^3.0.2",
